Product Create/Service
Commit made to save changes on files, the following additions were made:
Product Create service was created to keep Product Component clean as it is, the method is now sending information to the db.json file using BaseUrl and Http.post of HttpModule.

The create method is based on an observable<Product> that is called when createProduct() requests a .subscribe(() => xxx.